+++ playground/lib/pages/calendar.dart
@@ -1,5 +1,3 @@
-// ignore_for_file: avoid_print
-
 import 'package:flutter/material.dart';
 
 import 'package:shadcn_ui/shadcn_ui.dart';